update template and instructions
The default App Engine SA is no longer created automatically as part of
the Step 2 deployment, but as part of the manual step in Step 4 when
firestore database is created (where app engine application is created).
Therefore, the IAM assignment for that SA needs to be moved to Step 5 to
make sure the SA exists.